Configure push notifications in index.js
